switch from pnpm to yarn + nx, update db, write all graphql and rest modules

+++ apps/api/src/models/users/rest/users.controller.ts
@@ -9,8 +9,16 @@ import {
   Query,
 } from '@nestjs/common'
 
-import { ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger'
+import {
+  ApiBearerAuth,
+  ApiCreatedResponse,
+  ApiOkResponse,
+  ApiTags,
+} from '@nestjs/swagger'
+import { AllowAuthenticated, GetUser } from 'src/common/auth/auth.decorator'
+import { checkRowLevelPermission } from 'src/common/auth/util'
 import { PrismaService } from 'src/common/prisma/prisma.service'
+import { GetUserType } from 'src/common/types'
 import { CreateUser } from './dtos/create.dto'
 import { UserQueryDto } from './dtos/query.dto'
 import { UpdateUser } from './dtos/update.dto'
@@ -21,15 +29,12 @@ import { UserEntity } from './entity/user.entity'
 export class UsersController {
   constructor(private readonly prisma: PrismaService) {}
 
-  // @AllowAuthenticated()
-  // @ApiBearerAuth()
+  @AllowAuthenticated()
+  @ApiBearerAuth()
   @ApiCreatedResponse({ type: UserEntity })
   @Post()
-  create(
-    @Body() createUserDto: CreateUser,
-    // @GetUser() user: GetUserType
-  ) {
-    // checkRowLevelPermission(user, createUserDto.uid)
+  create(@Body() createUserDto: CreateUser, @GetUser() user: GetUserType) {
+    checkRowLevelPermission(user, createUserDto.uid)
     return this.prisma.user.create({ data: createUserDto })
   }
 
@@ -55,31 +60,28 @@ export class UsersController {
   }
 
   @ApiOkResponse({ type: UserEntity })
-  // @ApiBearerAuth()
-  // @AllowAuthenticated()
+  @ApiBearerAuth()
+  @AllowAuthenticated()
   @Patch(':uid')
   async update(
     @Param('uid') uid: string,
     @Body() updateUserDto: UpdateUser,
-    // @GetUser() user: GetUserType,
+    @GetUser() user: GetUserType,
   ) {
-    // const userInfo = await this.prisma.user.findUnique({ where: { uid } })
-    // checkRowLevelPermission(user, userInfo.uid)
+    const userInfo = await this.prisma.user.findUnique({ where: { uid } })
+    checkRowLevelPermission(user, userInfo.uid)
     return this.prisma.user.update({
       where: { uid },
       data: updateUserDto,
     })
   }
 
-  // @ApiBearerAuth()
-  // @AllowAuthenticated()
+  @ApiBearerAuth()
+  @AllowAuthenticated()
   @Delete(':uid')
-  async remove(
-    @Param('uid') uid: string,
-    //  @GetUser() user: GetUserType
-  ) {
-    // const userInfo = await this.prisma.user.findUnique({ where: { uid } })
-    // checkRowLevelPermission(user, userInfo.uid)
+  async remove(@Param('uid') uid: string, @GetUser() user: GetUserType) {
+    const userInfo = await this.prisma.user.findUnique({ where: { uid } })
+    checkRowLevelPermission(user, userInfo.uid)
     return this.prisma.user.delete({ where: { uid } })
   }
 }
